header and lists are own componets
