+++ src/views/Home.vue
@@ -17,172 +17,13 @@
         </ion-fab-button>
       </ion-fab>
 
-      <div
-        id="header"
-        slot="fixed"
-        style="background-color: var(--ion-color-secondary); width:100%; z-index:-10"
-      >
-        <ion-toolbar color="secondary" style="z-index: 0; margin-top: 32px">
-          <ion-title
-            color="tertiary"
-            style="color: var(--ion-color-tertiary);  position: static; line-height: 1.2"
-            size="large"
-            >Welcome <br />
-            on Board!</ion-title
-          >
-          <ion-text color="teritary">
-            <p
-              class="ion-margin-start"
-              style=" color: var(--ion-color-tertiary)"
-            >
-              Turn on Autoplay, and whenever you <br />
-              come across a point of interest, <br />
-              an audio track is played.
-            </p>
-          </ion-text>
-        </ion-toolbar>
-
-        <transition name="fade">
-          <div
-            class="ion-padding-horizontal"
-            style="background-color: var(--ion-color-secondary); margin-top: -24px; padding-bottom: 24px"
-            v-if="autoplay"
-          >
-            <ion-range
-              style="padding-inline: 0;"
-              v-model="playDistance"
-              @ionChange="setPlayDistance"
-              min="50"
-              max="2000"
-              step="40"
-              snaps="true"
-              ticks="false"
-              color="dark"
-              :pin="true"
-              ><ion-label slot="start">50M</ion-label
-              ><ion-label slot="end">2000M</ion-label></ion-range
-            >
-            <ion-note>
-              At how many meters distance should the trek be played.
-            </ion-note>
-          </div>
-        </transition>
-
-        <div
-          style="width: 80%; max-width: 380px; padding: -25px; margin-left: auto;  overflow: visible; margin-top: -64px"
-          id="landschaft"
-        ></div>
-        <div
-          style="color: var(--ion-color-danger); height: 100em; border-top: 1px solid #ccc; margin-top: -5px"
-        ></div>
-      </div>
-
-      <div
-        id="lists"
+      <Header
         v-if="initLoad"
-        style="z-index:10; background-color: var(--ion-toolbar-background); border-radius: 16px 16px 0 0; width: 100%; position: fixed; overflow: hidden; margin-top: 300px"
-      >
-        <ion-button
-          style="width: 100%; margin: 0"
-          shape="round"
-          fill="clear"
-          v-on:click="animateLists"
-          @drag="animteLists"
-        >
-          <ion-icon
-            style="width: 100%"
-            slot="icon-only"
-            :icon="listsIcon"
-          ></ion-icon>
-        </ion-button>
-
-        <ion-toolbar style="border-radius: 32px; margin-top: 0px">
-          <ion-text>
-            <h1 class="ion-margin-start" style="margin-top: 0px">
-              Places Nearby
-            </h1>
-          </ion-text>
-        </ion-toolbar>
-
-        <ion-grid>
-          <ion-row
-            class="ion-nowrap ion-align-items-start ion-padding-end"
-            style="overflow-x: scroll; scroll-snap-type: x mandatory; scroll-padding-left: 16px;"
-          >
-            <ion-col v-for="(page, index) in pages" :key="index">
-              <card-big
-                :img="page.mainImage.thumb"
-                @click-text="setAndOpenModal(index)"
-                @click-image="setAndPlayTrack(index)"
-              >
-                <template v-slot:subtitle>{{ page.dist + "M" }}</template>
-                <template v-slot:title>{{
-                  truncateAndClamps(page.title, 15, false)
-                }}</template>
-                <template v-slot:content>{{
-                  truncate(page.summary, 54, true)
-                }}</template>
-              </card-big>
-            </ion-col>
-          </ion-row>
-        </ion-grid>
-        <div
-          style="height: 100em; background-color: var(--ion-toolbar-background)"
-        ></div>
-        <!-- <ion-toolbar class="bottom-divider">
-              <ion-text>
-                <h1 class="ion-margin-start">Recently played</h1>
-              </ion-text>
-            </ion-toolbar>
-            <ion-grid>
-              <ion-row
-                class="ion-nowrap ion-align-items-start"
-                style="overflow-x: scroll;"
-              >
-                <ion-col v-for="(page, index) in pages" :key="index">
-                  <card-small
-                    :img="page.mainImage.thumb"
-                    @click-text="setAndOpenModal(page.pageID)"
-                    @click-image="setAndPlayTrack(page.pageID)"
-                  >
-                    <template v-slot:subtitle>{{ page.dist + "M" }}</template>
-                    <template v-slot:title>{{
-                      page.title.replace(/\(.*?\)/, "")
-                    }}</template>
-                    <template v-slot:content>{{
-                      truncate(page.summary, 36, true)
-                    }}</template>
-                  </card-small>
-                  <card-small
-                    :img="page.mainImage.thumb"
-                    @click-text="setAndOpenModal(page.pageID)"
-                    @click-image="setAndPlayTrack(page.pageID)"
-                  >
-                    <template v-slot:subtitle>{{ page.dist + "M" }}</template>
-                    <template v-slot:title>{{
-                      page.title.replace(/\(.*?\)/, "")
-                    }}</template>
-                    <template v-slot:content>{{
-                      truncate(page.summary, 36, true)
-                    }}</template>
-                  </card-small>
-                  <card-small
-                    :img="page.mainImage.thumb"
-                    @click-text="setAndOpenModal(page.pageID)"
-                    @click-image="setAndPlayTrack(page.pageID)"
-                  >
-                    <template v-slot:subtitle>{{ page.dist + "M" }}</template>
-                    <template v-slot:title>{{
-                      page.title.replace(/\(.*?\)/, "")
-                    }}</template>
-                    <template v-slot:content>{{
-                      truncate(page.summary, 36, true)
-                    }}</template>
-                  </card-small>
-                </ion-col>
-              </ion-row>
-            </ion-grid> -->
-      </div>
+        @animateLists="animateLists"
+        @setAndOpenModal="setAndOpenModal"
+        @setAndPlayTrack="setAndPlayTrack"
+      ></Header>
+      <Lists @animateLists="animateLists"></Lists>
     </ion-content>
 
     <ion-footer v-if="track">
@@ -209,7 +50,7 @@
           class="ion-margin-start"
           style="margin: 0 0 0 16px;"
         >
-          {{ trackTitle }}
+          {{ TextStore.truncate(track.title, 14, false) }}
         </h4>
       </ion-toolbar>
     </ion-footer>
@@ -253,10 +94,10 @@ import {
   IonNote,
 } from "@ionic/vue";
 
-import CardBig from "../components/CardBig.vue";
-import CardSmall from "../components/CardSmall.vue";
 import Modal from "../components/Modal.vue";
 import Player from "../components/Player.vue";
+import Header from "../components/Header.vue";
+import Lists from "../components/Lists.vue";
 
 import { defineComponent } from "vue";
 import { Plugins } from "@capacitor/core";
@@ -264,10 +105,9 @@ const { SplashScreen } = Plugins;
 
 import Store from "../store";
 import TrackStore from "../store/track.js";
+import TextStore from "../store/text.js";
 
 import gsap from "gsap";
-import lottie from "lottie-web";
-import landschaft from "../assets/landschaft.json";
 
 import {
   playCircleOutline,
@@ -276,8 +116,6 @@ import {
   playSkipBackOutline,
   playSkipForwardOutline,
   arrowBackOutline,
-  chevronUpOutline,
-  chevronDownOutline,
 } from "ionicons/icons";
 
 import { TextToSpeech } from "@ionic-native/text-to-speech";
@@ -287,10 +125,9 @@ export default defineComponent({
   data() {
     return {
       readingSpeed: "",
-      landschaft: null,
+
       modal: false,
       lists: { small: 54, big: null, current: 54 },
-      landschaftBottom: null,
       initLoad: false,
       pos: 0,
       playDistance: 2000,
@@ -316,14 +153,6 @@ export default defineComponent({
     }, 10000);
 
     SplashScreen.hide();
-
-    this.landschaft = lottie.loadAnimation({
-      container: document.getElementById("landschaft"), // the dom element that will contain the animation
-      renderer: "svg",
-      loop: true,
-      autoplay: true,
-      animationData: landschaft, // the path to the animation json
-    });
   },
   computed: {
     headerHeight() {
@@ -335,6 +164,9 @@ export default defineComponent({
     },
     TrackStore() {
       return TrackStore();
+    },
+    TextStore() {
+      return TextStore();
     },
     track() {
       return TrackStore().track.value;
@@ -349,24 +181,8 @@ export default defineComponent({
         return this.icons.playOutline;
       }
     },
-    listsIcon() {
-      if (this.lists.current <= this.lists.small) {
-        return chevronUpOutline;
-      } else {
-        return chevronDownOutline;
-      }
-    },
     audioPlaying() {
       return false;
-    },
-    pages() {
-      return Store().sortedPages.value;
-    },
-    position() {
-      return Store().position.value;
-    },
-    trackTitle() {
-      return this.truncateAndClamps(this.track.title, 14, false);
     },
   },
   methods: {
@@ -397,21 +213,7 @@ export default defineComponent({
         .then(() => console.log("Success Speach"))
         .catch((reason) => console.log(reason));
     },
-    truncate(str, n, useWordBoundary) {
-      if (str.length <= n) {
-        return str;
-      }
-      const subString = str.substr(0, n - 1); // the original check
-      return (
-        (useWordBoundary
-          ? subString.substr(0, subString.lastIndexOf(" "))
-          : subString) + "…"
-      );
-    },
-    truncateAndClamps(str, n, useWordBoundary) {
-      let string = str.replace(/\(.*?\)/, "");
-      return this.truncate(string, n, useWordBoundary);
-    },
+
     stopAndPlay(scrollPos) {
       if (scrollPos < 0) {
         scrollPos = 0;
@@ -420,16 +222,6 @@ export default defineComponent({
         scrollPos = 160;
       }
       //this.landschaft.goToAndStop(scrollPos * 25);
-    },
-    setAndOpenModal(index) {
-      this.TrackStore.setCurrentPageIndex(index);
-      this.modal = true;
-      console.log("page index", index);
-    },
-    setAndPlayTrack(index) {
-      this.TrackStore.setCurrentPageIndex(index);
-      this.TrackStore.play();
-      console.log("page index", index);
     },
     play() {
       this.TrackStore.play();
@@ -474,10 +266,10 @@ export default defineComponent({
     IonListHeader,
     IonList,
     IonModal,
-    CardBig,
-    CardSmall,
     Modal,
     Player,
+    Header,
+    Lists,
     IonSpinner,
     IonNote,
   },
